check user in group

+++ app/src/main/java/com/social/app/controller/CommentController.java
@@ -3,8 +3,12 @@ package com.social.app.controller;
 import com.social.app.entity.ResponseObject;
 import com.social.app.model.Comment;
 import com.social.app.model.Post;
+import com.social.app.model.User;
+import com.social.app.repository.GroupRepository;
+import com.social.app.repository.JoinRepository;
 import com.social.app.repository.PostRepository;
 import com.social.app.service.CommentService;
+import com.social.app.service.PostServices;
 import com.social.app.service.UserService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
@@ -26,6 +30,8 @@ public class CommentController {
     private CommentService commentService;
 
     @Autowired
+    private PostServices postServices;
+    @Autowired
     private PostRepository postRepository;
     @PostMapping("/{postID}/comments")
     @PreAuthorize("hasAuthority('ROLE_USER')")
@@ -33,6 +39,10 @@ public class CommentController {
                                                  @RequestParam("userid") int userid,
                                                  @PathVariable long postID){
         try {
+            // Get post from postId
+            Post post = postServices.loadPostById(postID);
+            // Check if user is not in group, user can not create comment
+            if(!userService.isGroupMember(userid, post.getGroup().getGroupId())) throw new RuntimeException("Must be group member");
             // set user for comment
             comment.setUser(userService.loadUserById(userid));
             // create comment
@@ -48,8 +58,13 @@ public class CommentController {
 
     @DeleteMapping("/delete/{commentID}")
     @PreAuthorize("hasAuthority('ROLE_USER')")
-    ResponseEntity<ResponseObject> deleteComment(@PathVariable long commentID){
+    ResponseEntity<ResponseObject> deleteComment(@PathVariable long commentID, @RequestParam("userid") int userid){
         try {
+            // Get comment from commentId
+            Comment comment = commentService.getCommentByID(commentID);
+            // Check if user is not in group, user can not delete
+            if(!userService.isGroupMember(userid, comment.getPost().getGroup().getGroupId())) throw new RuntimeException("Must be group member");
+            // Call delete method
             this.commentService.deleteComment(commentID);
             return ResponseEntity.status(HttpStatus.OK).body(new ResponseObject("Success", "Delete comment successfully", ""));
         }
@@ -68,8 +83,13 @@ public class CommentController {
 
     @PutMapping("/edit/{commentID}")
     @PreAuthorize("hasAuthority('ROLE_USER')")
-    ResponseEntity<ResponseObject> editComment(@PathVariable long commentID, @RequestBody Comment newComment){
+    ResponseEntity<ResponseObject> editComment(@PathVariable long commentID, @RequestPart Comment newComment, @RequestParam("userid") int userid){
         try {
+            // Get comment from commentId
+            Comment comment = commentService.getCommentByID(commentID);
+            // Check if user is not in group, user can not edit
+            if(!userService.isGroupMember(userid, comment.getPost().getGroup().getGroupId())) throw new RuntimeException("Must be group member");
+            // Call edit method
             Comment editedComment = this.commentService.editComment(newComment, commentID);
             return ResponseEntity.status(HttpStatus.OK).body(new ResponseObject("Success", "Edit comment successfully", editedComment));
         }
@@ -78,6 +98,8 @@ public class CommentController {
                     new ResponseObject("Failed", "There are problem..", ""));
         }
     }
+
+
